--- books.json
@@ -1,1 +1,0 @@
-[{"id":470,"publisher":"Ahmed","cover_state":"good","publish_date":"2019-02-02","label":{"title":"John is fine","color":"Green"}},{"id":663,"publisher":"Kevin","cover_state":"good","publish_date":"2016-04-04","label":{"title":"Something somehing","color":"blue"}},{"id":946,"publisher":"Ahmed Reda","cover_state":"good","publish_date":"2020-02-08","label":{"title":"I live in Cairo","color":"Blue"}}]
